+++ cordova.plugins.diagnostic.d.ts
@@ -25,6 +25,7 @@ interface Diagnostic {
         "GET_ACCOUNTS": "GET_ACCOUNTS";
         "ACCESS_FINE_LOCATION": "ACCESS_FINE_LOCATION";
         "ACCESS_COARSE_LOCATION": "ACCESS_COARSE_LOCATION";
+        "ACCESS_BACKGROUND_LOCATION": "ACCESS_BACKGROUND_LOCATION";
         "RECORD_AUDIO": "RECORD_AUDIO";
         "READ_PHONE_STATE": "READ_PHONE_STATE";
         "CALL_PHONE": "CALL_PHONE";
@@ -41,6 +42,7 @@ interface Diagnostic {
         "WRITE_EXTERNAL_STORAGE": "WRITE_EXTERNAL_STORAGE";
         "READ_EXTERNAL_STORAGE": "READ_EXTERNAL_STORAGE";
         "BODY_SENSORS": "BODY_SENSORS";
+        "ACTIVITY_RECOGNITION": "ACTIVITY_RECOGNITION";
     };
 
     /**
@@ -59,6 +61,7 @@ interface Diagnostic {
         "SENSORS": ["BODY_SENSORS"];
         "SMS": ["SEND_SMS", "RECEIVE_SMS", "READ_SMS", "RECEIVE_WAP_PUSH", "RECEIVE_MMS"];
         "STORAGE": ["READ_EXTERNAL_STORAGE", "WRITE_EXTERNAL_STORAGE"];
+        "PHYSICAL_ACTIVITY": ["ACTIVITY_RECOGNITION"];
     };
 
     /**
@@ -72,10 +75,10 @@ interface Diagnostic {
 
         // iOS only
         "RESTRICTED": "restricted";
-        "GRANTED_WHEN_IN_USE": "authorized_when_in_use";
 
         // Both iOS and Android
         "GRANTED": "authorized"|"GRANTED";
+        "GRANTED_WHEN_IN_USE": "authorized_when_in_use";
         "NOT_REQUESTED": "not_determined"|"NOT_REQUESTED";
         "DENIED_ALWAYS": "denied_always"|"DENIED_ALWAYS";
     };
@@ -96,13 +99,23 @@ interface Diagnostic {
     };
 
     /**
-     * iOS ONLY
+     * ANDROID and iOS ONLY
      * Location authorization mode
      * @type {Object}
      */
     locationAuthorizationMode: {
         "ALWAYS": "always";
         "WHEN_IN_USE": "when_in_use";
+    };
+
+    /**
+     * iOS ONLY
+     * Location accuracy authorization
+     * @type {Object}
+     */
+    locationAccuracyAuthorization: {
+        "FULL": "full";
+        "REDUCED": "reduced";
     };
 
 
@@ -340,12 +353,36 @@ interface Diagnostic {
      * Requests location authorization for the application.
      * @param successCallback
      * @param errorCallback
-     * @param mode - (iOS-only / optional) location authorization mode specified as a locationAuthorizationMode constant. If not specified, defaults to WHEN_IN_USE.
+     * @param mode - (optional / iOS & Android >= 10) location authorization mode specified as a locationAuthorizationMode constant. If not specified, defaults to WHEN_IN_USE.
      */
     requestLocationAuthorization?: (
         successCallback: (status: string) => void,
         errorCallback: (error: string) => void,
         mode?: string
+    ) => void;
+
+    /**
+     * iOS ONLY
+     * Returns the location accuracy authorization for the application.
+     * @param successCallback
+     * @param errorCallback
+     */
+    getLocationAccuracyAuthorization?: (
+        successCallback: (status: string) => void,
+        errorCallback: (error: string) => void
+    ) => void;
+
+    /**
+     * iOS ONLY
+     * Requests temporary access to full location accuracy for the application.
+     * @param purpose
+     * @param successCallback
+     * @param errorCallback
+     */
+    requestTemporaryFullAccuracyAuthorization?: (
+        purpose: string,
+        successCallback?: (status: string) => void,
+        errorCallback?: (error: string) => void
     ) => void;
 
 
@@ -502,7 +539,7 @@ interface Diagnostic {
      * ANDROID and iOS ONLY
      * Opens settings page for this app.
      * On Android, this opens the "App Info" page in the Settings app.
-     * On iOS, this opens the app settings page in the Settings app. This works only on iOS 8+ - iOS 7 and below will invoke the errorCallback.
+     * On iOS, this opens the app settings page in the Settings app.
      */
     switchToSettings?: (
         successCallback: () => void,
@@ -526,7 +563,7 @@ interface Diagnostic {
      * @param successCallback
      */
     registerBluetoothStateChangeHandler?: (
-        successCallback: (state: string) => void
+        successCallback?: (state: string|null) => void
     ) => void;
 
     /**
@@ -537,6 +574,15 @@ interface Diagnostic {
      * @param successCallback
      */
     registerLocationStateChangeHandler?: (
+        successCallback?: (state: string) => void
+    ) => void;
+
+    /**
+     * iOS ONLY
+     * Registers a function to be called when a change in location accuracy authorization occurs.
+     * @param successCallback
+     */
+    registerLocationAccuracyAuthorizationChangeHandler?: (
         successCallback: (state: string) => void
     ) => void;
 
@@ -565,6 +611,17 @@ interface Diagnostic {
      */
     getArchitecture?: (
         successCallback: (state: string) => void,
+        errorCallback: (error: string) => void
+    ) => void;
+
+    /**
+     * ANDROID and iOS ONLY
+     * Returns the current battery level of the device as a percentage.
+     * @param successCallback
+     * @param errorCallback
+     */
+    getCurrentBatteryLevel?: (
+        successCallback: (level: number) => void,
         errorCallback: (error: string) => void
     ) => void;
 
@@ -660,7 +717,7 @@ interface Diagnostic {
      * @param permissions
      */
     getPermissionsAuthorizationStatus?: (
-        successCallback: (status: string[]) => void,
+        successCallback: (status: Record<string,string>) => void,
         errorCallback: (error: string) => void,
         permissions: string[]
     ) => void;
@@ -686,7 +743,7 @@ interface Diagnostic {
      * @param permissions
      */
     requestRuntimePermissions?: (
-        successCallback: (status: string[]) => void,
+        successCallback: (status: Record<string,string>) => void,
         errorCallback: (error: string) => void,
         permissions: string[]
     ) => void;
@@ -703,7 +760,7 @@ interface Diagnostic {
      * @param successCallback
      */
     registerPermissionRequestCompleteHandler?: (
-        successCallback: (statuses: any) => void
+        successCallback?: (statuses: any) => void
     ) => void;
 
     /**
@@ -834,7 +891,7 @@ interface Diagnostic {
      * @param successCallback
      */
     registerNFCStateChangeHandler?: (
-        successCallback: (state: string) => void
+        successCallback?: (state: string) => void
     ) => void;
 
     /**