+++ www/android/diagnostic.js
@@ -37,6 +37,7 @@ var Diagnostic = (function(){
             "GET_ACCOUNTS": "GET_ACCOUNTS",
             "ACCESS_FINE_LOCATION": "ACCESS_FINE_LOCATION",
             "ACCESS_COARSE_LOCATION": "ACCESS_COARSE_LOCATION",
+            "ACCESS_BACKGROUND_LOCATION": "ACCESS_BACKGROUND_LOCATION",
             "RECORD_AUDIO": "RECORD_AUDIO",
             "READ_PHONE_STATE": "READ_PHONE_STATE",
             "CALL_PHONE": "CALL_PHONE",
@@ -52,7 +53,8 @@ var Diagnostic = (function(){
             "RECEIVE_MMS": "RECEIVE_MMS",
             "WRITE_EXTERNAL_STORAGE": "WRITE_EXTERNAL_STORAGE",
             "READ_EXTERNAL_STORAGE": "READ_EXTERNAL_STORAGE",
-            "BODY_SENSORS": "BODY_SENSORS"
+            "BODY_SENSORS": "BODY_SENSORS",
+            "ACTIVITY_RECOGNITION": "ACTIVITY_RECOGNITION"
         };
 
     /**
@@ -65,23 +67,35 @@ var Diagnostic = (function(){
             "CALENDAR": ["READ_CALENDAR", "WRITE_CALENDAR"],
             "CAMERA": ["CAMERA"],
             "CONTACTS": ["READ_CONTACTS", "WRITE_CONTACTS", "GET_ACCOUNTS"],
-            "LOCATION": ["ACCESS_FINE_LOCATION", "ACCESS_COARSE_LOCATION"],
+            "LOCATION": ["ACCESS_FINE_LOCATION", "ACCESS_COARSE_LOCATION", "ACCESS_BACKGROUND_LOCATION"],
             "MICROPHONE": ["RECORD_AUDIO"],
             "PHONE": ["READ_PHONE_STATE", "CALL_PHONE", "ADD_VOICEMAIL", "USE_SIP", "PROCESS_OUTGOING_CALLS", "READ_CALL_LOG", "WRITE_CALL_LOG"],
             "SENSORS": ["BODY_SENSORS"],
             "SMS": ["SEND_SMS", "RECEIVE_SMS", "READ_SMS", "RECEIVE_WAP_PUSH", "RECEIVE_MMS"],
-            "STORAGE": ["READ_EXTERNAL_STORAGE", "WRITE_EXTERNAL_STORAGE"]
+            "STORAGE": ["READ_EXTERNAL_STORAGE", "WRITE_EXTERNAL_STORAGE"],
+            "PHYSICAL_ACTIVITY": ["ACTIVITY_RECOGNITION"]
         };
 
     Diagnostic.runtimePermissionStatus = // deprecated
         Diagnostic.permissionStatus = {
-            "GRANTED": "GRANTED", //  User granted access to this permission, the device is running Android 5.x or below, or the app is built with API 22 or below.
-            "DENIED_ONCE": "DENIED_ONCE", // User denied access to this permission
-            "NOT_REQUESTED": "NOT_REQUESTED", // App has not yet requested access to this permission.
-            "DENIED_ALWAYS": "DENIED_ALWAYS" // User denied access to this permission and checked "Never Ask Again" box.
+            //  Location permission requested and 
+            //      app build SDK/user device is Android >10 and user granted background location ("all the time") permission,
+            //      or app build SDK/user device is Android 6-9 and user granted location permission,
+            //  or non-location permission requested
+            //      and app build SDK/user device is Android >=6 and user granted permission
+            //  or app build SDK/user device is Android <6
+            "GRANTED": "GRANTED",
+            //  Location permission requested 
+            //  and app build SDK/user device is Android >10 
+            //  and user granted background foreground location ("while-in-use") permission
+            "GRANTED_WHEN_IN_USE": "authorized_when_in_use",
+            // User denied access to this permission
+            "DENIED_ONCE": "DENIED_ONCE",
+            // User denied access to this permission and checked "Never Ask Again" box.
+            "DENIED_ALWAYS": "DENIED_ALWAYS",
+            // App has not yet requested access to this permission.
+            "NOT_REQUESTED": "NOT_REQUESTED"
         };
-
-
 
     Diagnostic.cpuArchitecture = {
         UNKNOWN: "unknown",
@@ -429,6 +443,23 @@ var Diagnostic = (function(){
             errorCallback,
             'Diagnostic',
             'isDataRoamingEnabled',
+            []);
+    };
+
+
+    /**
+     * Returns the current battery level of the device as a percentage.
+     *
+     * @param {Function} successCallback -  The callback which will be called when the operation is successful.
+     * This callback function is passed a single integer parameter which the current battery level percentage.
+     * @param {Function} errorCallback -  The callback which will be called when the operation encounters an error.
+     *  This callback function is passed a single string parameter containing the error message.
+     */
+    Diagnostic.getCurrentBatteryLevel = function(successCallback, errorCallback){
+        return cordova.exec(successCallback,
+            errorCallback,
+            'Diagnostic',
+            'getCurrentBatteryLevel',
             []);
     };
 
@@ -1025,6 +1056,17 @@ var Diagnostic = (function(){
             cordova.plugins.diagnostic.notifications.isRemoteNotificationsEnabled.apply(this, arguments);
         }else{
             throw "Diagnostic Notifications module is not installed";
+        }
+    };
+
+    /**
+     * Switches to the notification settings page in the Settings app
+     */
+    Diagnostic.switchToNotificationSettings = function() {
+        if (cordova.plugins.diagnostic.notifications){
+            cordova.plugins.diagnostic.notifications.switchToNotificationSettings.apply(this, arguments);
+        } else {
+            throw "Diagnostic notification module is not installed";
         }
     };